+++ uliweb/orm/__init__.py
@@ -1461,7 +1461,8 @@ class ReferenceProperty(Property):
         
         if __lazy_model_init__:
             if inspect.isclass(self.reference_class) and issubclass(self.reference_class, Model):
-                warnings.warn("Reference Model should be a string type, but [%s] model class found." % self.reference_class.__name__)
+                warnings.simplefilter('default')
+                warnings.warn("Reference Model should be a string type, but [%s] model class found." % self.reference_class.__name__, DeprecationWarning)
         
     def create(self, cls):
         global __nullable__